Renamed controller to dispatcher
